// src/additional-form/enums/air-force/air-force.type.enum.ts
export enum AirForceTypeEnum {
  GENERAL = '일반', // 일반
  ELECTRONIC_COMPUTER = '전자계산', // 전자계산
  CBR = '화생방', // 화생방
  MEDICAL = '의무', // 의무
  MECHANICAL = '기계', // 기계
  VEHICLE_DRIVING = '차량운전', // 차량운전
  VEHICLE_MAINTENANCE = '차량정비', // 차량정비
  COMMUNICATION_ELECTRONIC_ELECTRICAL = '통신전자전기', // 통신전자전기
}